requirements.txt
google-generativeai==0.3.2
python-dotenv==1.0.1
PyPDF2==3.0.1
tqdm==4.66.2
mistralai==0.0.12
tk==0.1.0  # Only needed if not included in Python installation